# specs/feature/multi-provider-ai/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ROXY JARVIS AI Provider API
  version: 1.0.0
  description: Multi-provider AI abstraction layer

servers:
  - url: /api/v1/ai

paths:
  /chat:
    post:
      operationId: chatCompletion
      summary: Send a chat completion request
      tags: [AI]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIRequest'
      responses:
        '200':
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIResponse'
            text/event-stream:
              schema:
                type: string
                example: ": provider=deepseek model=deepseek-chat-v3\ndata: {\"delta\":\"Hello\",\"done\":false}\ndata: {\"delta\":\"!\",\"done\":false}\ndata: [DONE]"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          description: All providers unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /providers:
    get:
      operationId: listProviders
      summary: List all configured providers and their status
      tags: [AI]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderHealthResponse'

  /providers/{provider_name}:
    patch:
      operationId: updateProvider
      summary: Enable or disable a provider (operator override)
      tags: [AI]
      security:
        - bearerAuth: []
      parameters:
        - name: provider_name
          in: path
          required: true
          schema:
            type: string
            enum: [deepseek, grok, openai, gemini]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
      responses:
        '200':
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderStatus'
        '404':
          $ref: '#/components/responses/NotFound'

  /providers/health:
    get:
      operationId: providerHealth
      summary: Health check for all providers (no auth required)
      tags: [AI]
      responses:
        '200':
          description: Health status of all providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderHealthResponse'

  /usage:
    get:
      operationId: getTokenUsage
      summary: Get recent token usage logs
      tags: [AI]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
        - name: provider
          in: query
          schema:
            type: string
            enum: [deepseek, grok, openai, gemini]
      responses:
        '200':
          description: Token usage logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  logs:
                    type: array
                    items:
                      $ref: '#/components/schemas/TokenUsageLog'
                  total_cost_usd:
                    type: number

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    AIRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        provider:
          type: string | null
          description: Request-level provider override
        task_type:
          type: string
          enum: [general, coding, reasoning]
          default: general
        temperature:
          type: number
          default: 0.7
          minimum: 0
          maximum: 2
        max_tokens:
          type: integer | null
          minimum: 1
        stream:
          type: boolean
          default: false
        timeout_seconds:
          type: number
          minimum: 0.1
          maximum: 60

    Message:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
        content:
          type: string

    AIResponse:
      type: object
      properties:
        content:
          type: string
        provider:
          type: string
        model:
          type: string
        agent_id:
          type: string
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cost_usd:
          type: number
        latency_ms:
          type: integer
        request_id:
          type: string
          format: uuid

    ProviderStatus:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [healthy, degraded, unavailable]
        circuit_state:
          type: string
          enum: [closed, half_open, open]
        last_error:
          type: string | null
        last_success:
          type: string | null
          format: date-time

    ProviderHealthResponse:
      type: object
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderStatus'

    TokenUsageLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        provider:
          type: string
        model:
          type: string
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cost_usd:
          type: number
        latency_ms:
          type: integer
        timestamp:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        request_id:
          type: string
          format: uuid

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'